requirements.txt
Flask==2.1.1
Flask-PyMongo==2.3.0
pymongo==4.8.0
Werkzeug==2.1.2
gunicorn==23.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
requests>=2.25.0
dnspython>=2.0.0
